Working LLDBPlugin implemented
* Very basic implementation of LLDBPlugin.

* Starts hardcoded executable, sets a breakpoint at "main" and exits when the running process has stopped.

+++ testbed/plugins/LLDB/LLDBPlugin.cpp
@@ -6,6 +6,9 @@
 #include <SBListener.h>
 #include <SBProcess.h>
 #include <SBDebugger.h>
+#include <SBHostOS.h>
+#include <SBEvent.h>
+#include <SBBreakpoint.h>
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -13,6 +16,7 @@ typedef struct LLDBPlugin
 {
 	lldb::SBDebugger debugger;
 	lldb::SBTarget target;
+	lldb::SBListener listener;
 
 } LLDBPlugin;
 
@@ -20,6 +24,7 @@ typedef struct LLDBPlugin
 
 static void* createInstance(ServiceFunc* serviceFunc)
 {
+	printf("Create instance\n");
 	LLDBPlugin* plugin = new LLDBPlugin; 
 	return plugin;
 }
@@ -77,13 +82,78 @@ static void startCallback(void* userData, PDLaunchAction action, void* launchDat
 {
 	LLDBPlugin* plugin = (LLDBPlugin*)userData;
 
+	printf("Trying to start debug session\n");
+
+	lldb::SBDebugger::Initialize();
+	lldb::SBHostOS::ThreadCreated ("<lldb-tester.app.main>");
+ 	plugin->debugger = lldb::SBDebugger::Create(false);
+ 	plugin->listener = plugin->debugger.GetListener(); 
+
+	const char* cat[] =
+	{
+		"all",
+		0
+	};
+
 	// very hacky right now but is only used for testing anyway
 	plugin->target = plugin->debugger.CreateTarget((const char*)launchData);
-	lldb::SBProcess process = plugin->target.LaunchSimple(0, 0, 0);
-	usleep(10000);
-	process.Kill();
-	process.Destroy();
-	lldb::SBDebugger::Destroy(plugin->debugger);
+	lldb::SBBreakpoint breakpoint = plugin->target.BreakpointCreateByName("main");
+
+	if (breakpoint.IsValid())
+	{
+		printf("Added valid breakpoint\n");
+	}
+
+	if (!plugin->debugger.EnableLog("lldb", cat))
+	{
+		printf("Unable to enable log!\n");
+	}
+	
+	if (!plugin->target.IsValid())
+		return;
+
+	printf("Target is valid, launching\n");
+
+	lldb::SBLaunchInfo launchInfo(0);
+	lldb::SBError error;
+
+	lldb::SBProcess process = plugin->target.Launch(launchInfo, error);
+
+	if (!error.Success())
+	{
+		printf("error false\n");
+		return;
+	}
+
+	if (!process.IsValid())
+	{
+		printf("process not valid\n");
+		return;
+	}
+
+	process.GetBroadcaster().AddListener(
+			plugin->listener, 
+			lldb::SBProcess::eBroadcastBitStateChanged | lldb::SBProcess::eBroadcastBitInterrupt);
+
+    lldb::SBEvent evt;
+
+    while (1)
+	{
+		plugin->listener.WaitForEvent (UINT32_MAX, evt);
+		lldb::StateType state =  lldb::SBProcess::GetStateFromEvent (evt);
+		printf("event = %s\n", lldb::SBDebugger::StateAsCString(state));
+
+		if (state == lldb::eStateStopped)
+		{
+			process.Kill();
+			return;
+		}
+
+		usleep(10000);
+	}
+
+	//process.Destroy();
+	//lldb::SBDebugger::Destroy(plugin->debugger);
 }
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -104,7 +174,8 @@ extern "C"
 
 void InitPlugin(int version, ServiceFunc* serviceFunc, RegisterPlugin* registerPlugin)
 {
-	registerPlugin(PD_DEBUGGER_TYPE, &plugin);
+	printf("Starting to register Plugin!\n");
+	registerPlugin(PD_PTYPE_DEBUGGER, &plugin);
 }
 
 }
